requirements.txt
# Core
numpy<2.0.0
protobuf==3.20.3
pandas>=2.0.0
tqdm
click

# Media Processing
ffmpeg-python
# opencv-python-headless  <-- REMOVED to avoid conflict with mediapipe's opencv-contrib-python
librosa
soundfile

# AI & Vision
mediapipe==0.10.11
faster-whisper

# App & Viz
streamlit
altair
matplotlib

# Dev
pytest
black
flake8
